benchmarks/requirements.txt
# Benchmark dependencies
Pillow>=9.0.0
exifread>=2.3.0
psutil>=5.8.0

